# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Access Mobile Virtual Mining API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RegisterRequest {
  username: string;
  password: string;
  referralCode?: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export interface User {
  id: number;
  username: string;
  balance: number;
  referralCode: string;
  referralCount: number;
  avatarUrl?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface UpdateAvatarRequest {
  /** Data URI (e.g. data:image/jpeg;base64,...) of the new avatar image. */
  avatarBase64: string;
}

export interface MiningStatus {
  isActive: boolean;
  startedAt: string | null;
  /** Current server time, used by the client to keep the countdown in sync and immune to local clock changes. */
  serverNow: string;
  /** Total duration of a mining session in milliseconds (12 hours). */
  sessionDurationMs: number;
  /** Milliseconds remaining until the active session can be claimed. 0 when no session is active or it is ready. */
  remainingMs: number;
  /** True once the active session has run its full duration and can be claimed. */
  isClaimable: boolean;
  /** Projected (not yet credited) amount accrued so far in the active session, for display only. */
  earnedSoFar: number;
  ratePerSession: number;
  balance: number;
  activeReferralCount: number;
  bonusPerSession: number;
  totalMined: number;
  totalSupply: number;
}

export interface ClaimResponse {
  earned: number;
  newBalance: number;
  activeReferralCount: number;
}

export interface LeaderboardEntry {
  rank: number;
  username: string;
  balance: number;
  avatarUrl?: string | null;
}

export interface LeaderboardResponse {
  entries: LeaderboardEntry[];
  totalMined: number;
  totalSupply: number;
}

export interface ReferralEntry {
  id: number;
  username: string;
  avatarUrl?: string | null;
  balance: number;
  isActive: boolean;
  joinedAt: string;
}

export interface ReferralsResponse {
  referrals: ReferralEntry[];
  totalCount: number;
  referralCode: string;
}

export interface Transaction {
  id: number;
  amount: number;
  type: string;
  createdAt: string;
}

export interface TransactionsResponse {
  transactions: Transaction[];
}