# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Access Mobile Virtual Mining API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: auth
  - name: mining
  - name: leaderboard
  - name: referrals
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /mining/status:
    get:
      operationId: getMiningStatus
      tags: [mining]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Mining status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MiningStatus"

  /mining/start:
    post:
      operationId: startMining
      tags: [mining]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Mining started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MiningStatus"
        "400":
          description: Already mining
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /mining/claim:
    post:
      operationId: claimMining
      tags: [mining]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Tokens claimed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClaimResponse"
        "400":
          description: No active session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/me/avatar:
    patch:
      operationId: updateAvatar
      tags: [auth]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAvatarRequest"
      responses:
        "200":
          description: Avatar updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /referrals:
    get:
      operationId: getReferrals
      tags: [referrals]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Your referrals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferralsResponse"

  /leaderboard:
    get:
      operationId: getLeaderboard
      tags: [leaderboard]
      responses:
        "200":
          description: Leaderboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderboardResponse"

  /transactions:
    get:
      operationId: getTransactions
      tags: [mining]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionsResponse"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    RegisterRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        referralCode:
          type: string
      required: [username, password]

    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]

    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]

    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        balance:
          type: number
        referralCode:
          type: string
        referralCount:
          type: integer
        avatarUrl:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, username, balance, referralCode, referralCount, createdAt]

    UpdateAvatarRequest:
      type: object
      properties:
        avatarBase64:
          type: string
          description: Data URI (e.g. data:image/jpeg;base64,...) of the new avatar image.
      required: [avatarBase64]

    MiningStatus:
      type: object
      properties:
        isActive:
          type: boolean
        startedAt:
          type: string
          nullable: true
        serverNow:
          type: string
          description: Current server time, used by the client to keep the countdown in sync and immune to local clock changes.
        sessionDurationMs:
          type: number
          description: Total duration of a mining session in milliseconds (12 hours).
        remainingMs:
          type: number
          description: Milliseconds remaining until the active session can be claimed. 0 when no session is active or it is ready.
        isClaimable:
          type: boolean
          description: True once the active session has run its full duration and can be claimed.
        earnedSoFar:
          type: number
          description: Projected (not yet credited) amount accrued so far in the active session, for display only.
        ratePerSession:
          type: number
        balance:
          type: number
        activeReferralCount:
          type: integer
        bonusPerSession:
          type: number
        totalMined:
          type: number
        totalSupply:
          type: number
      required: [isActive, startedAt, serverNow, sessionDurationMs, remainingMs, isClaimable, earnedSoFar, ratePerSession, balance, activeReferralCount, bonusPerSession, totalMined, totalSupply]

    ClaimResponse:
      type: object
      properties:
        earned:
          type: number
        newBalance:
          type: number
        activeReferralCount:
          type: integer
      required: [earned, newBalance, activeReferralCount]

    LeaderboardEntry:
      type: object
      properties:
        rank:
          type: integer
        username:
          type: string
        balance:
          type: number
        avatarUrl:
          type: string
          nullable: true
      required: [rank, username, balance]

    LeaderboardResponse:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/LeaderboardEntry"
        totalMined:
          type: number
        totalSupply:
          type: number
      required: [entries, totalMined, totalSupply]

    ReferralEntry:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        avatarUrl:
          type: string
          nullable: true
        balance:
          type: number
        isActive:
          type: boolean
        joinedAt:
          type: string
      required: [id, username, balance, isActive, joinedAt]

    ReferralsResponse:
      type: object
      properties:
        referrals:
          type: array
          items:
            $ref: "#/components/schemas/ReferralEntry"
        totalCount:
          type: integer
        referralCode:
          type: string
      required: [referrals, totalCount, referralCode]

    Transaction:
      type: object
      properties:
        id:
          type: integer
        amount:
          type: number
        type:
          type: string
        createdAt:
          type: string
      required: [id, amount, type, createdAt]

    TransactionsResponse:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
      required: [transactions]