# lib/api-zod/src/generated/types/referralEntry.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Access Mobile Virtual Mining API
 * OpenAPI spec version: 0.1.0
 */

export interface ReferralEntry {
  id: number;
  username: string;
  avatarUrl?: string | null;
  balance: number;
  isActive: boolean;
  joinedAt: string;
}