# lib/api-zod/src/generated/types/referralsResponse.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Access Mobile Virtual Mining API
 * OpenAPI spec version: 0.1.0
 */
import type { ReferralEntry } from './referralEntry';

export interface ReferralsResponse {
  referrals: ReferralEntry[];
  totalCount: number;
  referralCode: string;
}