# lib/api-zod/src/generated/types/updateAvatarRequest.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Access Mobile Virtual Mining API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateAvatarRequest {
  /** Data URI (e.g. data:image/jpeg;base64,...) of the new avatar image. */
  avatarBase64: string;
}